Add "application" concept and refactor (#5)
This introduces the concept of an "application" and `ApplicationState`
struct to represent the current state of a migration -- one of
pending, applied, changed or removed. We can then refactor the cli
`main` function quite a bit to operate over a vector of
`ApplicationState`.

To track whether a migration has been changed, or whether a migration
definition has been removed, we add up/down string columns to the
migrations table and store what these sql strings were at the time of
application. This will eventually allow us to revert old or removed
migrations.

Also adds integration tests.